requirements.txt
# Video Generator Dependencies
edge-tts>=6.1.0
moviepy>=1.0.3
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
Pillow>=9.0.0 